Workaround for mod ores
yes, all this does is generate a regular stone, instead of the stone from this mod.
And yes, ores from other mods are not generated in this stone.
This mod should complement the caves, so the fact that ores are generated only in the overworld is incorrect, they should also be generated in the dimensions from this mod.